kernel tests

+++ parallelCUDA/kernelTests.cu.h
@@ -194,6 +194,7 @@ int setSecondMatrixTest(float* d_A, float* d_B, int dim1, int dim2, int batchsiz
 
     { // timing the GPU implementations
         gettimeofday(&t_start, NULL);
+        
         for(int i=0; i<RUNS_GPU; i++) {
             int numBlocks = (batchsize * dim1 * dim1 + BLOCKSIZE - 1) / BLOCKSIZE;
             setSecondMatrix<<<numBlocks, BLOCKSIZE>>>(d_PointerA, d_PointerB, d_n1Tilde, d_n1Union, d_n2, dim1, batchsize);
@@ -202,7 +203,7 @@ int setSecondMatrixTest(float* d_A, float* d_B, int dim1, int dim2, int batchsiz
         gettimeofday(&t_end, NULL);
         timeval_subtract(&t_diff, &t_end, &t_start);
         elapsed = (t_diff.tv_sec*1e6+t_diff.tv_usec) / RUNS_GPU;
-        gigaBytesPerSec = 2 * dim1 * dim1 * batchsize * sizeof(int) * 1.0e-3f / elapsed;
+        gigaBytesPerSec = 2 * dim1 * dim1 * batchsize * sizeof(float) * 1.0e-3f / elapsed;
         printf("\n\nParallel secondMatrix runs in: %lu microsecs, GB/sec: %.2f\n\n\n"
               , elapsed, gigaBytesPerSec);
     }
